+++ components/datetime/src/pattern/hour_cycle.rs
@@ -57,8 +57,8 @@ impl CoarseHourCycle {
     #[cfg(feature = "provider_transform_internals")]
     pub fn apply_on_pattern<'data>(
         &self,
-        date_time: &provider::gregory::patterns::GenericLengthPatternsV1<'data>,
-        skeletons: &provider::gregory::DateSkeletonPatternsV1<'data>,
+        date_time: &provider::calendar::patterns::GenericLengthPatternsV1<'data>,
+        skeletons: &provider::calendar::DateSkeletonPatternsV1<'data>,
         pattern_str: &str,
         mut pattern: reference::Pattern,
     ) -> Option<reference::Pattern> {
